sql: separate mysql and postgresql clients into different Go files
