+++ pkg/clients/database/mysql.go
@@ -25,8 +25,6 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/services/mysql/mgmt/2017-12-01/mysql"
 	"github.com/Azure/azure-sdk-for-go/services/mysql/mgmt/2017-12-01/mysql/mysqlapi"
-	"github.com/Azure/azure-sdk-for-go/services/postgresql/mgmt/2017-12-01/postgresql"
-	"github.com/Azure/azure-sdk-for-go/services/postgresql/mgmt/2017-12-01/postgresql/postgresqlapi"
 	"github.com/Azure/go-autorest/autorest/azure/auth"
 	"github.com/pkg/errors"
 
@@ -36,7 +34,6 @@ import (
 	azure "github.com/crossplaneio/stack-azure/pkg/clients"
 )
 
-// Helper functions
 // NOTE: postgresql and mysql structs and functions live in their respective
 // packages even though they are exactly the same. However, Crossplane does not
 // make that assumption and use the respective package for each type, although,
@@ -264,192 +261,6 @@ func UpdateMySQLVirtualNetworkRuleStatusFromAzure(v *azuredbv1alpha3.MySQLServer
 	v.Status.Type = azure.ToString(az.Type)
 }
 
-// PostgreSQLServerAPI represents the API interface for a MySQL Server client
-type PostgreSQLServerAPI interface {
-	ServerNameTaken(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer) (bool, error)
-	GetServer(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer) (postgresql.Server, error)
-	CreateServer(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer, adminPassword string) error
-	DeleteServer(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer) error
-	UpdateServer(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer) error
-}
-
-// PostgreSQLServerClient is the concreate implementation of the SQLServerAPI interface for PostgreSQL that calls Azure API.
-type PostgreSQLServerClient struct {
-	postgresql.ServersClient
-	postgresql.CheckNameAvailabilityClient
-}
-
-// NewPostgreSQLServerClient creates and initializes a PostgreSQLServerClient instance.
-func NewPostgreSQLServerClient(c *azure.Client) (*PostgreSQLServerClient, error) {
-	postgreSQLServerClient := postgresql.NewServersClient(c.SubscriptionID)
-	postgreSQLServerClient.Authorizer = c.Authorizer
-	postgreSQLServerClient.AddToUserAgent(azure.UserAgent)
-
-	nameClient := postgresql.NewCheckNameAvailabilityClient(c.SubscriptionID)
-	nameClient.Authorizer = c.Authorizer
-	nameClient.AddToUserAgent(azure.UserAgent)
-
-	return &PostgreSQLServerClient{
-		ServersClient:               postgreSQLServerClient,
-		CheckNameAvailabilityClient: nameClient,
-	}, nil
-}
-
-// ServerNameTaken returns true if the supplied server's name has been taken.
-func (c *PostgreSQLServerClient) ServerNameTaken(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer) (bool, error) {
-	r, err := c.Execute(ctx, postgresql.NameAvailabilityRequest{Name: azure.ToStringPtr(meta.GetExternalName(s))})
-	if err != nil {
-		return false, err
-	}
-	return !azure.ToBool(r.NameAvailable), nil
-}
-
-// GetServer retrieves the requested PostgreSQL Server
-func (c *PostgreSQLServerClient) GetServer(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer) (postgresql.Server, error) {
-	return c.ServersClient.Get(ctx, s.Spec.ForProvider.ResourceGroupName, meta.GetExternalName(s))
-}
-
-// CreateServer creates a PostgreSQL Server
-func (c *PostgreSQLServerClient) CreateServer(ctx context.Context, cr *azuredbv1alpha3.PostgreSQLServer, adminPassword string) error {
-	s := cr.Spec.ForProvider
-	properties := &postgresql.ServerPropertiesForDefaultCreate{
-		AdministratorLogin:         azure.ToStringPtr(s.AdministratorLogin),
-		AdministratorLoginPassword: &adminPassword,
-		Version:                    postgresql.ServerVersion(s.Version),
-		SslEnforcement:             postgresql.SslEnforcementEnum(s.SSLEnforcement),
-		CreateMode:                 postgresql.CreateModeDefault,
-		StorageProfile: &postgresql.StorageProfile{
-			BackupRetentionDays: azure.ToInt32PtrFromIntPtr(s.StorageProfile.BackupRetentionDays),
-			GeoRedundantBackup:  postgresql.GeoRedundantBackup(azure.ToString(s.StorageProfile.GeoRedundantBackup)),
-			StorageMB:           azure.ToInt32Ptr(s.StorageProfile.StorageMB),
-			StorageAutogrow:     postgresql.StorageAutogrow(azure.ToString(s.StorageProfile.StorageAutogrow)),
-		},
-	}
-	sku, err := ToPostgreSQLSKU(s.SKU)
-	if err != nil {
-		return err
-	}
-	createParams := postgresql.ServerForCreate{
-		Sku:        sku,
-		Properties: properties,
-		Location:   &s.Location,
-		Tags:       azure.ToStringPtrMap(s.Tags),
-	}
-	op, err := c.Create(ctx, s.ResourceGroupName, meta.GetExternalName(cr), createParams)
-	if err != nil {
-		return err
-	}
-	_, err = op.DoneWithContext(ctx, c.ServersClient.Client)
-	return err
-}
-
-// UpdateServer updates a PostgreSQL Server.
-func (c *PostgreSQLServerClient) UpdateServer(ctx context.Context, cr *azuredbv1alpha3.PostgreSQLServer) error {
-	// TODO(muvaf): password update via Update call is supported by Azure but
-	// we don't support that.
-	s := cr.Spec.ForProvider
-	properties := &postgresql.ServerUpdateParametersProperties{
-		Version:        postgresql.ServerVersion(s.Version),
-		SslEnforcement: postgresql.SslEnforcementEnum(s.SSLEnforcement),
-		StorageProfile: &postgresql.StorageProfile{
-			BackupRetentionDays: azure.ToInt32PtrFromIntPtr(s.StorageProfile.BackupRetentionDays),
-			GeoRedundantBackup:  postgresql.GeoRedundantBackup(azure.ToString(s.StorageProfile.GeoRedundantBackup)),
-			StorageMB:           azure.ToInt32Ptr(s.StorageProfile.StorageMB),
-			StorageAutogrow:     postgresql.StorageAutogrow(azure.ToString(s.StorageProfile.StorageAutogrow)),
-		},
-	}
-	sku, err := ToPostgreSQLSKU(s.SKU)
-	if err != nil {
-		return err
-	}
-	updateParams := postgresql.ServerUpdateParameters{
-		Sku:                              sku,
-		ServerUpdateParametersProperties: properties,
-		Tags:                             azure.ToStringPtrMap(s.Tags),
-	}
-	op, err := c.Update(ctx, s.ResourceGroupName, meta.GetExternalName(cr), updateParams)
-	if err != nil {
-		return err
-	}
-	_, err = op.DoneWithContext(ctx, c.ServersClient.Client)
-	return err
-}
-
-// DeleteServer deletes the given PostgreSQL resource
-func (c *PostgreSQLServerClient) DeleteServer(ctx context.Context, s *azuredbv1alpha3.PostgreSQLServer) error {
-	op, err := c.ServersClient.Delete(ctx, s.Spec.ForProvider.ResourceGroupName, meta.GetExternalName(s))
-	if err != nil {
-		return err
-	}
-	_, err = op.DoneWithContext(ctx, c.ServersClient.Client)
-	return err
-}
-
-// A PostgreSQLVirtualNetworkRulesClient handles CRUD operations for Azure Virtual Network Rules.
-type PostgreSQLVirtualNetworkRulesClient postgresqlapi.VirtualNetworkRulesClientAPI
-
-// NewPostgreSQLVirtualNetworkRulesClient returns a new Azure Virtual Network Rules client. Credentials must be
-// passed as JSON encoded data.
-func NewPostgreSQLVirtualNetworkRulesClient(ctx context.Context, credentials []byte) (PostgreSQLVirtualNetworkRulesClient, error) {
-	c := azure.Credentials{}
-	if err := json.Unmarshal(credentials, &c); err != nil {
-		return nil, errors.Wrap(err, "cannot unmarshal Azure client secret data")
-	}
-
-	client := postgresql.NewVirtualNetworkRulesClient(c.SubscriptionID)
-
-	cfg := auth.ClientCredentialsConfig{
-		ClientID:     c.ClientID,
-		ClientSecret: c.ClientSecret,
-		TenantID:     c.TenantID,
-		AADEndpoint:  c.ActiveDirectoryEndpointURL,
-		Resource:     c.ResourceManagerEndpointURL,
-	}
-	a, err := cfg.Authorizer()
-	if err != nil {
-		return nil, errors.Wrapf(err, "cannot create Azure authorizer from credentials config")
-	}
-	client.Authorizer = a
-	if err := client.AddToUserAgent(azure.UserAgent); err != nil {
-		return nil, errors.Wrap(err, "cannot add to Azure client user agent")
-	}
-
-	return client, nil
-}
-
-// NewPostgreSQLVirtualNetworkRuleParameters returns an Azure VirtualNetworkRule object from a virtual network spec
-func NewPostgreSQLVirtualNetworkRuleParameters(v *azuredbv1alpha3.PostgreSQLServerVirtualNetworkRule) postgresql.VirtualNetworkRule {
-	return postgresql.VirtualNetworkRule{
-		Name: azure.ToStringPtr(v.Spec.Name),
-		VirtualNetworkRuleProperties: &postgresql.VirtualNetworkRuleProperties{
-			VirtualNetworkSubnetID:           azure.ToStringPtr(v.Spec.VirtualNetworkRuleProperties.VirtualNetworkSubnetID),
-			IgnoreMissingVnetServiceEndpoint: azure.ToBoolPtr(v.Spec.VirtualNetworkRuleProperties.IgnoreMissingVnetServiceEndpoint, azure.FieldRequired),
-		},
-	}
-}
-
-// PostgreSQLServerVirtualNetworkRuleNeedsUpdate determines if a virtual network rule needs to be updated
-func PostgreSQLServerVirtualNetworkRuleNeedsUpdate(kube *azuredbv1alpha3.PostgreSQLServerVirtualNetworkRule, az postgresql.VirtualNetworkRule) bool {
-	up := NewPostgreSQLVirtualNetworkRuleParameters(kube)
-
-	switch {
-	case !reflect.DeepEqual(up.VirtualNetworkRuleProperties.VirtualNetworkSubnetID, az.VirtualNetworkRuleProperties.VirtualNetworkSubnetID):
-		return true
-	case !reflect.DeepEqual(up.VirtualNetworkRuleProperties.IgnoreMissingVnetServiceEndpoint, az.VirtualNetworkRuleProperties.IgnoreMissingVnetServiceEndpoint):
-		return true
-	}
-
-	return false
-}
-
-// UpdatePostgreSQLVirtualNetworkRuleStatusFromAzure updates the status related to the external
-// Azure PostgreSQLVirtualNetworkRule in the VirtualNetworkStatus
-func UpdatePostgreSQLVirtualNetworkRuleStatusFromAzure(v *azuredbv1alpha3.PostgreSQLServerVirtualNetworkRule, az postgresql.VirtualNetworkRule) {
-	v.Status.State = string(az.VirtualNetworkRuleProperties.State)
-	v.Status.ID = azure.ToString(az.ID)
-	v.Status.Type = azure.ToString(az.Type)
-}
-
 // The name must match the specification of the SKU, so, we don't allow user
 // to specify an arbitrary name. The format is tier + family + cores, e.g. B_Gen4_1, GP_Gen5_8.
 
@@ -466,33 +277,6 @@ func ToMySQLSKU(skuSpec azuredbv1alpha3.SKU) (*mysql.Sku, error) {
 		Family:   azure.ToStringPtr(skuSpec.Family),
 		Size:     skuSpec.Size,
 	}, nil
-}
-
-// ToPostgreSQLSKU returns a *postgresql.Sku object that can be used in Azure API calls.
-func ToPostgreSQLSKU(skuSpec azuredbv1alpha3.SKU) (*postgresql.Sku, error) {
-	t, ok := skuShortTiers[mysql.SkuTier(skuSpec.Tier)]
-	if !ok {
-		return nil, fmt.Errorf("tier '%s' is not one of the supported values: %+v", skuSpec.Tier, mysql.PossibleSkuTierValues())
-	}
-	return &postgresql.Sku{
-		Name:     azure.ToStringPtr(fmt.Sprintf("%s_%s_%s", t, skuSpec.Family, strconv.Itoa(skuSpec.Capacity))),
-		Tier:     postgresql.SkuTier(skuSpec.Tier),
-		Capacity: azure.ToInt32Ptr(skuSpec.Capacity),
-		Family:   azure.ToStringPtr(skuSpec.Family),
-		Size:     skuSpec.Size,
-	}, nil
-}
-
-// GeneratePostgreSQLObservation produces SQLServerObservation from postgresql.Server.
-func GeneratePostgreSQLObservation(in postgresql.Server) azuredbv1alpha3.SQLServerObservation {
-	return azuredbv1alpha3.SQLServerObservation{
-		ID:                       azure.ToString(in.ID),
-		Name:                     azure.ToString(in.Name),
-		Type:                     azure.ToString(in.Type),
-		UserVisibleState:         string(in.UserVisibleState),
-		FullyQualifiedDomainName: azure.ToString(in.FullyQualifiedDomainName),
-		MasterServerID:           azure.ToString(in.MasterServerID),
-	}
 }
 
 // GenerateMySQLObservation produces SQLServerObservation from mysql.Server
@@ -524,51 +308,6 @@ func LateInitializeMySQL(p *azuredbv1alpha3.SQLServerParameters, in mysql.Server
 // IsMySQLUpToDate is used to report whether given mysql.Server is in
 // sync with the SQLServerParameters that user desires.
 func IsMySQLUpToDate(p azuredbv1alpha3.SQLServerParameters, in mysql.Server) bool { // nolint:gocyclo
-	if in.StorageProfile == nil || in.Sku == nil {
-		return false
-	}
-	switch {
-	case p.SSLEnforcement != string(in.SslEnforcement):
-		return false
-	case p.Version != string(in.Version):
-		return false
-	case !reflect.DeepEqual(azure.ToStringPtrMap(p.Tags), in.Tags):
-		return false
-	case p.SKU.Tier != string(in.Sku.Tier):
-		return false
-	case p.SKU.Capacity != azure.ToInt(in.Sku.Capacity):
-		return false
-	case p.SKU.Family != azure.ToString(in.Sku.Family):
-		return false
-	case !reflect.DeepEqual(azure.ToInt32PtrFromIntPtr(p.StorageProfile.BackupRetentionDays), in.StorageProfile.BackupRetentionDays):
-		return false
-	case azure.ToString(p.StorageProfile.GeoRedundantBackup) != string(in.StorageProfile.GeoRedundantBackup):
-		return false
-	case p.StorageProfile.StorageMB != azure.ToInt(in.StorageProfile.StorageMB):
-		return false
-	case azure.ToString(p.StorageProfile.StorageAutogrow) != string(in.StorageProfile.StorageAutogrow):
-		return false
-	}
-	return true
-}
-
-// LateInitializePostgreSQL fills the empty values of SQLServerParameters with the
-// ones that are retrieved from the Azure API.
-func LateInitializePostgreSQL(p *azuredbv1alpha3.SQLServerParameters, in postgresql.Server) {
-	if in.Sku != nil {
-		p.SKU.Size = azure.LateInitializeStringPtrFromPtr(p.SKU.Size, in.Sku.Size)
-	}
-	p.Tags = azure.LateInitializeStringMap(p.Tags, in.Tags)
-	if in.StorageProfile != nil {
-		p.StorageProfile.BackupRetentionDays = azure.LateInitializeIntPtrFromInt32Ptr(p.StorageProfile.BackupRetentionDays, in.StorageProfile.BackupRetentionDays)
-		p.StorageProfile.GeoRedundantBackup = azure.LateInitializeStringPtrFromVal(p.StorageProfile.GeoRedundantBackup, string(in.StorageProfile.GeoRedundantBackup))
-		p.StorageProfile.StorageAutogrow = azure.LateInitializeStringPtrFromVal(p.StorageProfile.StorageAutogrow, string(in.StorageProfile.StorageAutogrow))
-	}
-}
-
-// IsPostgreSQLUpToDate is used to report whether given postgresql.Server is in
-// sync with the SQLServerParameters that user desires.
-func IsPostgreSQLUpToDate(p azuredbv1alpha3.SQLServerParameters, in postgresql.Server) bool { // nolint:gocyclo
 	if in.StorageProfile == nil || in.Sku == nil {
 		return false
 	}
